Move Entity mappings to SQLContext project
