+++ D.MG/Models/Debt.cs
@@ -1,18 +1,7 @@
-﻿using Microsoft.EntityFrameworkCore.Metadata.Builders;
-
-namespace Models
+﻿namespace Models
 {
     public class Debt : EntityBase
     {
 
     }
-
-    public class DebtMap
-    {
-        public DebtMap(EntityTypeBuilder<Debt> entityBuilder)
-        {
-            entityBuilder.HasKey(t => t.Id);
-        }
-    }
-
 }
