+++ D.MG/Models/Payment.cs
@@ -1,17 +1,7 @@
-﻿using Microsoft.EntityFrameworkCore.Metadata.Builders;
-
-namespace Models
+﻿namespace Models
 {
     public class Payment : EntityBase
     {
 
-    }
-
-    public class PaymentMap
-    {
-        public PaymentMap(EntityTypeBuilder<Payment> entityBuilder)
-        {
-            entityBuilder.HasKey(t => t.Id);
-        }
     }
 }
