+++ D.MG/Models/Settlement.cs
@@ -1,17 +1,7 @@
-﻿using Microsoft.EntityFrameworkCore.Metadata.Builders;
-
-namespace Models
+﻿namespace Models
 {
     public class Settlement : EntityBase
     {
 
-    }
-
-    public class SettlementMap
-    {
-        public SettlementMap(EntityTypeBuilder<Settlement> entityBuilder)
-        {
-            entityBuilder.HasKey(t => t.Id);
-        }
     }
 }
