+++ D.MG/Models/User.cs
@@ -1,9 +1,4 @@
-﻿using Microsoft.EntityFrameworkCore.Metadata.Builders;
-using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Models
+﻿namespace Models
 {
     public class User : EntityBase
     {
@@ -13,13 +8,5 @@ namespace Models
         public string Password { get; set; }
         public string County { get; set; }
         public string Address { get; set; }
-    }
-
-    public class UserMap
-    {
-        public UserMap(EntityTypeBuilder<User> entityBuilder)
-        {
-            entityBuilder.HasKey(t => t.Id);
-        }
     }
 }
